requirements-dev.txt
# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
coverage==7.2.7
pre-commit==3.3.3
python-decouple==3.8
python-magic==0.4.27
psutil==5.9.5
gunicorn==21.2.0
